Marty Site 0.11
\\ Service Area Changes //

— Changes made to the table.

— Instead of using 3 different tables for every county, we are now
using a single table with every area in an alphabetical order.

\\ Known Issues //

— Image on the lading page is overlapping with the header. This is
occurring in every browsers except for Internet Explorer and Microsoft
Edge.

— There is no submit button on the Contact Us form. This will be
implemented when wee have a basic server setup to send the data to.

— The Service Area list is hard for the user to read or find if their
area is on the list. We will be implementing a search function for
users to quickly and easily find if their area is on the list at a
later date. Until then Ctrl + F (Windows) or Cmd + F (Mac) works
perfectly fine.